# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Perguntas de Conexão API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/quiz/email-config:
    get:
      operationId: getAdminQuizEmailConfig
      tags: [admin]
      summary: Get the quiz email capture configuration
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Quiz email capture configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizEmailConfig"
        "403":
          description: Access denied
    patch:
      operationId: updateAdminQuizEmailConfig
      tags: [admin]
      summary: Update whether quiz email capture is required
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizEmailConfigInput"
      responses:
        "200":
          description: Updated quiz email capture configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizEmailConfig"
        "400":
          description: Invalid configuration
        "403":
          description: Access denied
...
  /admin/landing-pages/primary:
    get:
      operationId: getAdminPrimaryLandingPage
      tags: [admin]
      summary: Get the landing page configured for the main route
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current primary landing page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrimaryLandingPageResponse"
        "403":
          description: Access denied
    patch:
      operationId: updateAdminPrimaryLandingPage
      tags: [admin]
      summary: Change the landing page configured for the main route
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrimaryLandingPageInput"
      responses:
        "200":
          description: Primary landing page updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrimaryLandingPageResponse"
        "400":
          description: Invalid landing page
        "403":
          description: Access denied
  /admin/landing-pages/visibility:
    patch:
      operationId: updateAdminLandingPageVisibility
      tags: [admin]
      summary: Show or hide a landing page
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandingPageVisibilityInput"
      responses:
        "200":
          description: Updated landing page visibility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandingPageVisibility"
        "400":
          description: Invalid visibility change
        "403":
          description: Access denied
...
  /admin/session-recording:
    get:
      operationId: getAdminSessionRecording
      tags: [admin]
      summary: Find a Clarity session recording for a buyer
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
        - name: buyerId
          in: query
          required: false
          schema: { type: string }
        - name: visitorKey
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Recording lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionRecordingLookup"
        "403":
          description: Access denied
  /admin/lp-sessions:
    get:
      operationId: listAdminLpSessions
      tags: [admin]
      summary: List recent sessions for a landing page
      parameters:
        - name: sessionId
          in: query
          required: true
          schema: { type: string }
        - name: lpId
          in: query
          required: true
          schema:
            type: string
            enum: [v1, v2, lp3]
      responses:
        "200":
          description: Recent landing page sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/LandingSession"
                required: [sessions]
        "403":
          description: Access denied
  /push/vapid-public-key:
    get:
      operationId: getPushVapidPublicKey
      tags: [admin]
      summary: Return the public key used by Web Push
      responses:
        "200":
          description: VAPID public key
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey: { type: string }
                required: [publicKey]
        "503":
          description: Push notifications are not configured
  /admin/push-subscribe:
    post:
      operationId: subscribeAdminPush
      tags: [admin]
      summary: Register an administrator device for push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId: { type: string }
                subscription:
                  type: object
                  properties:
                    endpoint: { type: string }
                    keys:
                      type: object
                      properties:
                        p256dh: { type: string }
                        auth: { type: string }
                      required: [p256dh, auth]
                  required: [endpoint, keys]
              required: [sessionId, subscription]
      responses:
        "201":
          description: Subscription registered
        "400":
          description: Invalid subscription
        "403":
          description: Access denied
  /admin/push-unsubscribe:
    post:
      operationId: unsubscribeAdminPush
      tags: [admin]
      summary: Remove an administrator device from push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId: { type: string }
                endpoint: { type: string }
              required: [sessionId, endpoint]
      responses:
        "200":
          description: Subscription removed
        "400":
          description: Invalid endpoint
        "403":
          description: Access denied
components:
  schemas:
    PrimaryLandingPageId:
      type: string
      enum: [v1, v2, lp3]
    PrimaryLandingPageInput:
      type: object
      properties:
        primaryLandingPage:
          $ref: "#/components/schemas/PrimaryLandingPageId"
      required: [primaryLandingPage]
    PrimaryLandingPageResponse:
      type: object
      properties:
        primaryLandingPage:
          $ref: "#/components/schemas/PrimaryLandingPageId"
        fallbackUsed: { type: boolean }
        landingPages:
          type: array
          items:
            $ref: "#/components/schemas/LandingPageVisibility"
      required: [primaryLandingPage, fallbackUsed, landingPages]
    LandingPageVisibility:
      type: object
      properties:
        landingPage:
          $ref: "#/components/schemas/PrimaryLandingPageId"
        visible: { type: boolean }
      required: [landingPage, visible]
    LandingPageVisibilityInput:
      type: object
      properties:
        landingPage:
          $ref: "#/components/schemas/PrimaryLandingPageId"
        visible: { type: boolean }
      required: [landingPage, visible]
...
    LandingSession:
      type: object
      properties:
        visitorKey: { type: string }
        firstSeenAt: { type: string, format: date-time }
        timeOnPageSeconds: { type: ["number", "null"] }
        lastSection: { type: ["string", "null"] }
        status:
          type: string
          enum: [comprou, aguardando_pagamento, so_visitou]
        buyerName: { type: ["string", "null"] }
        packageName: { type: ["string", "null"] }
        hasRecording: { type: boolean }
      required:
        [
          visitorKey,
          firstSeenAt,
          timeOnPageSeconds,
          lastSection,
          status,
          buyerName,
          packageName,
          hasRecording,
        ]
...
    QuizEmailConfig:
      type: object
      properties:
        required: { type: boolean }
        liveSince: { type: ["string", "null"], format: date-time }
      required: [required, liveSince]
    QuizEmailConfigInput:
      type: object
      properties:
        required: { type: boolean }
      required: [required]
...
    SessionRecordingLookup:
      type: object
      properties:
        available: { type: boolean }
        url: { type: string }
        visitorKey: { type: string }
        reason:
          type: string
          enum: [sem-rastreio, sem-gravacao]
      required: [available]
...