# lib/api-zod/src/generated/types/quizEmailConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perguntas de Conexão API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizEmailConfig {
  required: boolean;
  /** @nullable */
  liveSince: Date | null;
}